rag/requirements.txt
# Core API Framework
fastapi==0.120.0
uvicorn==0.38.0
python-dotenv==1.1.1

# Vector Database & Embeddings
chromadb==1.2.1

# LangChain Integration
langchain-chroma==1.0.0
langchain-openai==1.0.1
